% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{ABM}
\alias{ABM}
\title{Effective emissivity from atmosphere with cloud atenuation}
\usage{
ABM(data, func, coef1 = 0.3, adjust = FALSE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}
}
\value{
a vector with emissivity estimatives
}
\description{
Effective emissivity from atmosphere with cloud atenuation
}
\author{
Roilan Hernandez, Guilherme Goergen and Jonatan Tatsch
}
\references{
Stöckli R (2007) LBA-MIP driver data gap filling algorithms.
Unpublished. http://www.climatemodeling.org/lba-mip/LBAmipDriverDataFillingMethods.pdf
Accessed 7 May 2011
}

