% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{ALH}
\alias{ALH}
\title{Effective emissivity from atmosphere with cloud atenuation}
\usage{
ALH(data, func, coef1 = 1.18, coef2 = 1.24, coef3 = -0.34, coef4 = 1.37,
  adjust = FALSE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1, coef2, coef3, coef4}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}
}
\value{
a vector with emissivity estimatives
}
\description{
Effective emissivity from atmosphere with cloud atenuation
}
\references{
Lhomme JP, Vacher JJ, Rocheteau A (2007) Estimating downward
long-wave radiation on the Andean Altiplano. Agr For Meteorol
145:139–148
}

