% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2eval.R
\name{CalcStats}
\alias{CalcStats}
\title{Function for statistical analise}
\usage{
CalcStats(data_li, statistic = c("rmse", "pbias"))
}
\arguments{
\item{data_li}{Data frame with observed and calculated  downward longwave radiation series.}

\item{statistic}{Statistical index to be calculated between observation and each simulation, "rmse" default.}
}
\description{
Recommended use of function from hydroGOF package (ex., rmse, mae, pbias, me, nrmse, rSD...)
}
\examples{
LiStats <- CalcStats(data_li = data_li, statistic = "pbias")
LiStats
}

