% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{CCB}
\alias{CCB}
\title{Amount of cloud estimatives functions.}
\usage{
CCB(data)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}
}
\value{
a vector with cloud amount estimatives
}
\description{
Amount of cloud estimatives functions.
}
\references{
Campbell, G.S. (1985) Soil physics with BASIC. Elsevier, Amsterdam
}

