% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{CFG}
\alias{CFG}
\title{Amount of cloud estimatives functions.}
\usage{
CFG(data)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}
}
\value{
a vector with cloud amount estimatives
}
\description{
Amount of cloud estimatives functions.
}
\references{
Flerchinger, G. N. (2009) Comparison of algoritmhs for incoming 
atmospheric long-wave radiation, Water Resources Res., 45, W0342.
}
