% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{CJG}
\alias{CJG}
\title{Amount of cloud estimatives functions.}
\usage{
CJG(data)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}
}
\value{
a vector with cloud amount estimatives
}
\description{
Amount of cloud estimatives functions.
}
\author{
Roilan Hernandez, Guilherme Goergen and Jonatan Tatsch
}

