% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{CKC}
\alias{CKC}
\title{Amount of cloud estimatives functions.}
\usage{
CKC(data)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}
}
\value{
a vector with cloud amount estimatives
}
\description{
Amount of cloud estimatives functions.
}
\references{
Kasten F, Czeplak G (1980) Solar and terrestrial radiation dependent on the amount and type of cloud. Sol Energy 24:177–189
}

