% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{CKZ}
\alias{CKZ}
\title{Amount of cloud estimatives functions.}
\usage{
CKZ(data, alt = 88)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{alt}{Site sea level heigth}
}
\value{
a vector with cloud amount estimatives
}
\description{
Amount of cloud estimatives functions.
}
\references{
Konzelmann T, van de Wal RSW, Greuell W, Bintanja R, Henneken  EAC, Abe-Ouchi A 
(1994) Parameterization of global and longwave incoming radiation for the Greenland Ice Sheet. 
Glob Planet Chang 9:143–164
}
