% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{CLM}
\alias{CLM}
\title{Amount of cloud estimatives functions.}
\usage{
CLM(data)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}
}
\value{
a vector with cloud amount estimatives
}
\description{
Amount of cloud estimatives functions.
}
\references{
Stockli R (2007) LBA-MIP driver data gap filling algorithms.
Unpublished. http://www.climatemodeling.org/lba-mip/LBAmipDriverDataFillingMethods.pdf
Accessed 7 May 2011
}
