% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.desc.R
\docType{data}
\name{data2}
\alias{data2}
\title{Santa Maria micrometeorological observations}
\format{\describe{
\item{date}{Date in POSIX format.}
\item{Rg}{Global radiation( W m-2)}
\item{Ta}{Air Temperature (K)}
\item{rh}{Relative humidity.}
\item{Li}{Downward Longwave Radiation (W m-2).}
\item{Rpot}{Potencial Radiation (W m-2).}
\item{es}{Water vapor partial pressure (Pa)}
\item{K}{Cloudness atenuation index (-)}
}}
\usage{
data2
}
\description{
Santa Maria micrometeorological observations
}
\keyword{datasets}
