% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.liexamp.R
\docType{data}
\name{data_li}
\alias{data_li}
\title{Downward longwave radiation observed and calculated}
\format{\describe{
\item{Li}{Downward longwave radiation observed}
\item{EBR}{Downward longwave radiation calculated by Brunt (1932) scheme}
\item{EKZ}{Downward longwave radiation calculated by Konzelmann (1994) scheme}
\item{FAN_CQB}{Downward longwave radiation calculated by Angstrom (1915) combined with the Quadratic regression of Black (1956) scheme}
}}
\usage{
data_li
}
\description{
Downward longwave radiation observed and calculated
}
\keyword{datasets}
