% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{EAI}
\alias{EAI}
\title{Emissivity from atmosphere}
\usage{
EAI(data, func, coef1 = 0.52843, coef2 = -0.0082, coef3 = 24242.6501,
  adjust = FALSE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1, coef2, coef3}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}
}
\value{
a vector with emissivity estimatives
}
\description{
Emissivity from atmosphere
}
\references{
Aimi, D. (2017); TODO
}

