% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{EAN}
\alias{EAN}
\title{Emissivity from atmosphere}
\usage{
EAN(data, func = "-", coef1 = 0.83, coef2 = 0.18, coef3 = 0.067,
  coef4 = 0.22, coef5 = 1, adjust = FALSE, forced = TRUE)
}
\arguments{
\item{data}{Data frame with all atmospherics variables}

\item{func}{Function for amount of cloud}

\item{coef1, coef2, coef3, coef4, coef5}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}

\item{forced}{Forced adjust, default TRUE}
}
\value{
a vector with emissivity estimatives
}
\description{
Emissivity from atmosphere
}
\references{
Angstrom, A. (1915) A study of the radiation of the atmosphere.
Smithsonian Miscellaneous Collections 65(3)
}
