% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{EBT}
\alias{EBT}
\title{Emissivity from atmosphere}
\usage{
EBT(data, func = "-", coef1 = 1.24, coef2 = 1/7, coef3 = 0.22,
  coef4 = 1, adjust = FALSE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1, coef2, coef3, coef4}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}
}
\value{
a vector with emissivity estimatives
}
\description{
Emissivity from atmosphere
}
\references{
Brutsaert W (1975) On a derivable formula for long-wave radiation
from clear skies. Water Resour Res 11:742–744
}

