% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{EID}
\alias{EID}
\title{Emissivity from atmosphere}
\usage{
EID(data, func, coef1 = 0.7, coef2 = 5.95e-05, adjust = FALSE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1, coef2}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}
}
\value{
a vector with emissivity estimatives
}
\description{
Emissivity from atmosphere
}

