% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{EIJ}
\alias{EIJ}
\title{Emissivity from atmosphere}
\usage{
EIJ(data, func = "-", coef1 = 0.261, coef2 = 0.000777, coef3 = 0.22,
  coef4 = 1, adjust = FALSE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1, coef2, coef3, coef4}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}
}
\value{
a vector with emissivity estimatives
}
\description{
Emissivity from atmosphere
}
\references{
Idso SB, Jackson RD (1969) Thermal radiation from the atmosphere.
J Geophys Res 74:5397–5403
}

