% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{EST}
\alias{EST}
\title{Emissivity from atmosphere}
\usage{
EST(data, func = "-", coef1 = 1.08, coef3 = 0.22, coef4 = 1,
  adjust = FALSE, forced = TRUE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1, coef3, coef4}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}

\item{forced}{Forced adjust, default TRUE}
}
\value{
a vector with emissivity estimatives
}
\description{
Emissivity from atmosphere
}
\references{
Sutterlund, D. R. (1979) An improved equation for estimating longwave 
radiation from the atmosphere, Water Res. Res., 15, 1649-1650.
}
