% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{ESW}
\alias{ESW}
\title{Emissivity from atmosphere}
\usage{
ESW(data, func = "-", coef1 = 9.2e-06, coef2 = 0.22, coef3 = 1,
  adjust = FALSE, forced = TRUE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1, coef2, coef3}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}

\item{forced}{Forced adjust, default TRUE}
}
\value{
a vector with emissivity estimatives
}
\description{
Emissivity from atmosphere
}
\references{
Swinbank WC (1963) Long-wave radiation from clear skies. Q J Roy
Meteor Soc 89:339–348
}
