% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{FIJ}
\alias{FIJ}
\title{Effective emissivity from atmosphere with cloud atenuation}
\usage{
FIJ(data, func, coef1 = 0.22, adjust = FALSE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}
}
\value{
a vector with emissivity estimatives
}
\description{
Effective emissivity from atmosphere with cloud atenuation
}
\references{
Idso SB, Jackson RD (1969) Thermal radiation from the atmosphere. 
J Geophys Res 74:5397–5403
}

