% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{FKZ}
\alias{FKZ}
\title{Effective emissivity from atmosphere with cloud atenuation}
\usage{
FKZ(data, func, coef1 = 0.963, adjust = FALSE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}
}
\value{
a vector with emissivity estimatives
}
\description{
Effective emissivity from atmosphere with cloud atenuation
}
\references{
Konzelmann T, van de Wal RSW, Greuell W, Bintanja R, Henneken
EAC, Abe-Ouchi A (1994) Parameterization of global and
longwave incoming radiation for the Greenland Ice Sheet. Glob
Planet Chang 9:143–164
}

