% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{kloudines}
\alias{kloudines}
\title{Incoming solar radiation atenuattion (K)}
\usage{
kloudines(Rg, lon = -53.76, lat = -29.72, timezone = -4)
}
\arguments{
\item{Rg}{global radiation serie.}

\item{lon}{longitude from local analisys}

\item{lat}{latitude from local analisys}
}
\value{
Data frame with date and K index columns
}
\description{
Incoming solar radiation atenuattion (K)
}
\author{
Roilan Hernandez, Guilherme Goergen and Jonatan Tatsch
}

