% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFuns.R
\name{LHSU}
\alias{LHSU}
\title{Create a LHS sequence from a interval for each parameter}
\usage{
LHSU(xmin, xmax, nsample)
}
\arguments{
\item{xmin}{qkbf}

\item{xmax}{ADKVN}

\item{nsample}{SLKDJhvb}
}
\description{
Create a LHS sequence from a interval for each parameter
}

