% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFuns.R
\name{maxlim}
\alias{maxlim}
\title{Limits vector values between interval}
\usage{
maxlim(i, max_ = 1, min_ = 0)
}
\arguments{
\item{i}{vector}

\item{max_}{Maximum value}

\item{min_}{Minimum value}
}
\description{
Limits vector values between interval
}
