% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2eval.R
\name{MonteCarlo}
\alias{MonteCarlo}
\title{MonteCarlo simulation for optimization}
\usage{
MonteCarlo(data, E_fun, func, coefs, nsample, max_iter, stats)
}
\arguments{
\item{data}{Data frame with all atmospherics variables}

\item{E_fun}{Function for emissivity}

\item{func}{Function from cloud cover calculation}

\item{coefs}{Coeficients in E_fun function}

\item{nsample}{Number of population individuous}

\item{max_iter}{Maximun number of iterations}

\item{stats}{Statistical index to be minimized}
}
\value{
Vector with best combination of parameters
}
\description{
Make a MonteCarlo simulation from initial value on each schemes' coeficients. Initial sample 
has a 'nsample'*10 population number for each parameter in scheme, determined by a LHS 
(Latin Hypercubic Sample) with frontiers +- 5 times the value of coeficient:
$$ A_i = [a_i-5a_i;a_i+5a_i] $$
The MonteCarlo design follows a similar way in genetic algorithms, where a initial population 
of value for  parameter produce a L_i serie that is evaluated with choosen estatistic (stats).
Always two equal dimenson population are compared and choosen combinations of parameters that 
produce a 'stats' minor than a o.5 quantile. That way ensures that best combinations be conserved
if its performance is better than a half of total combinations 'stats' value. Finally a coeficients 
combination with best performance prevales and is considered the optimization. Furthermore, 
a 'nsample', 'max_iter' and 'stats' will define a result.
}

