% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2eval.R
\name{PotRad}
\alias{PotRad}
\title{Potencial Radiation from date vector}
\usage{
PotRad(date, lon = -53.76, lat = -29.72, timezone = -4)
}
\arguments{
\item{date}{A vector with data}

\item{lon}{Longitude from local analisys}

\item{lat}{Latitude from local analisys}

\item{timezone}{Local time diference with GMT (-1 for fluxes measurement)}
}
\value{
Vector with potential radiation for given dates and coordinates.
}
\description{
Potencial Radiation from date vector
}
\author{
Roilan Hernandez
}

