% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2eval.R
\name{R2}
\alias{R2}
\title{Linear Regression Coeficient}
\usage{
R2(sim, obs, na.rm = TRUE)
}
\arguments{
\item{sim}{Simulation}

\item{obs}{Observation}

\item{na.rm}{preferible TRUE}
}
\description{
Linear Regression Coeficient
}

