% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2eval.R
\name{run_fun}
\alias{run_fun}
\title{Calculate emissivity}
\usage{
run_fun(E_fun, data, func, new.coefs)
}
\arguments{
\item{E_fun}{qkbf}

\item{data}{ADKVN}

\item{func}{SLKDJhvb}

\item{new.coefs}{qeohv}
}
\description{
Calculate emissivity
}
