% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2eval.R
\name{run_stats}
\alias{run_stats}
\title{Calculate statistics}
\usage{
run_stats(stats, obs, sim, ...)
}
\arguments{
\item{stats}{qkbf}

\item{obs}{ADKVN}

\item{sim}{SLKDJhvb}
}
\description{
Calculate statistics
}

