% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2eval.R
\name{split_stats}
\alias{split_stats}
\title{Split data and Li series by time factor (ex., "daytime", "season", "cover")}
\usage{
split_stats(data_, split_class = c("daytime", "season", "Cover", "ID"),
  lon = -53.18, lat = -29.71, timezone = -3, round = 3)
}
\arguments{
\item{data_}{Data frame with column date (POSIXct), Obs and all series for each scheme.}

\item{split_class}{Type of cut data analises (daytime and season by default, other types needs as a factor column)}

\item{lat, lon, timezone}{Latitude, longitude e timezone of observations local.}

\item{round}{Digits for round}
}
\description{
Split data and Li series by time factor (ex., "daytime", "season", "cover")
}
