% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2eval.R
\name{table_stats}
\alias{table_stats}
\title{Calculate statistical index}
\usage{
table_stats(data_, statistic = "rmse", round = 2)
}
\arguments{
\item{data_}{Data frame with column Li observed and all series for each scheme.}

\item{statistic}{Statistical index to calculate}

\item{round}{Digits for round}
}
\description{
Calculate statistical index
}
