% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2eval.R
\name{to.daylight}
\alias{to.daylight}
\title{Function for day and nigth identification}
\usage{
to.daylight(date, lon = -47.63, lat = -21.61, timezone = -4)
}
\arguments{
\item{date}{A vector with dates}

\item{lon}{Longitude from local analisys}

\item{lat}{Latitude from local analisys}

\item{timezone}{Local time diference with GMT (-1 for fluxes measurement)}
}
\value{
Vector with "day"/"night" string
}
\description{
Function for day and nigth identification
}
\author{
Roilan Hernandez
}

