% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{CKZ}
\alias{CKZ}
\title{Amount of cloud estimatives functions.}
\usage{
CKZ(data, alt = 88)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{alt}{Site sea level heigth}
}
\value{
a vector with cloud amount estimatives
}
\description{
Amount of cloud estimatives functions.
}
\author{
Roilan Hernandez, Guilherme Goergen and Jonatan Tatsch
}

