% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{CQB}
\alias{CQB}
\title{Amount of cloud estimatives functions.}
\usage{
CQB(data)
}
\arguments{
\item{x}{a numeric vector with the meteorological station code}
}
\value{
a data frame with variables in columns and observations along rows
}
\description{
Amount of cloud estimatives functions.
}
\details{
A minimum quality control check is applied to the data
This include: a chronological sequence check; filling missing dates with NA; 
remove duplicated data; aggregate time and date information into a POSIX object
}
\author{
Roilan Hernandez, Guilherme Goergen and Jonatan Tatsch
}

