% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{CQB}
\alias{CQB}
\title{Amount of cloud estimatives functions.}
\usage{
CQB(data)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}
}
\value{
a vector with cloud amount estimatives
}
\description{
Amount of cloud estimatives functions.
}
\references{
Black JN (1956) The distribution of solar radiation over the Earth's surface.
Arch Meteor Geophy B 7:165–189
}

