% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{CWU}
\alias{CWU}
\title{Amount of cloud estimatives functions.}
\usage{
CWU(data)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}
}
\value{
a vector with cloud amount estimatives
}
\description{
Amount of cloud estimatives functions.
}
\references{
Weishampel JF, Urban DL (1996) Coupling a spatially-explicit forest
gap model with a 3-D solar routine to simulate latitudinal effects.
Ecol Model 86:101–111
}
