% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{EBR}
\alias{EBR}
\title{Emissivity from atmosphere}
\usage{
EBR(data, func, coef1 = 0.51, coef2 = 0.066, adjust = FALSE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1, coef2}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}
}
\value{
a vector with emissivity estimatives
}
\description{
Emissivity from atmosphere
}
\references{
Brunt D (1932) Notes on radiation in the atmosphere I. Q J Roy
Meteor Soc 58:389–420
}

