% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{EDO}
\alias{EDO}
\title{Emissivity from atmosphere}
\usage{
EDO(data, func = "-", coef1 = 59.38, coef2 = 113.7, coef3 = 96.96,
  coef4 = 0.22, coef5 = 1, adjust = FALSE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1, coef2, coef3, coef4, coef5}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}
}
\value{
a vector with emissivity estimatives
}
\description{
Emissivity from atmosphere
}
\references{
Dilley , A. C. (1998) Estimating downward clear-sky long-wave 
irradiance at the surface from screen temperature and precpitable water. 
Q. J. R. Meteorol. Soc., 96, 313-319.
}

