% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{EGB}
\alias{EGB}
\title{Effective emissivity from atmosphere with cloud atenuation}
\usage{
EGB(data, func = "-", coef1 = 0.84, coef2 = 21, coef3 = 0.22,
  coef4 = 1, adjust = FALSE, method = "non-linear", nsample = 100,
  max_iter = 10, stats = "rmse")
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1, coef2, coef3, coef4}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}

\item{method}{"non-linear" (default) for Non linear Least Square adjust, 
"montecarlo" for MonteCarlo optimization. Later be usseful when a NLS can't 
adjust observed data allowing optimization.}

\item{nsample}{population number evaluated in each iteration 
(only when method = "montecarlo").}

\item{max_iter}{maximun number of iterations (only when method = "montecarlo").}

\item{stats}{statistical function to be minimized (only when method = "montecarlo"),
NOTE: the best result should be 0.0 (ex., if stats = r (correlation), then transform to 
rMod = 1.0 - r, so the best result is when r== 1.0, so rMod == 0.0)}
}
\value{
a vector with emissivity estimatives
}
\description{
Effective emissivity from atmosphere with cloud atenuation
}
\references{
Gabathuler M, Marty CA, Hanselmann KW (2001) Parameterization
of incoming longwave radiation in high-mountain environments.
Phys Geogr 22,99–114
}

