% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{EGR}
\alias{EGR}
\title{Emissivity from atmosphere}
\usage{
EGR(data, func = "-", coef1 = 0.79, coef2 = 0.17, coef3 = 0.096,
  coef4 = 0.22, coef5 = 1, adjust = FALSE, forced = TRUE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1, coef2, coef3, coef4, coef5}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}

\item{forced}{Forced adjust, default TRUE}
}
\value{
a vector with emissivity estimatives
}
\description{
Emissivity from atmosphere
}
\references{
Garratt, J.A. (1992), Extreme maximun land surface temperatures,
J. Appl. Meteorol., 31, 1096-1105.
}
