% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{EID}
\alias{EID}
\title{Emissivity from atmosphere}
\usage{
EID(data, func = "-", coef1 = 0.7, coef2 = 5.95e-05, coef3 = 0.22,
  coef4 = 1, adjust = FALSE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1, coef2, coef3, coef4}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}
}
\value{
a vector with emissivity estimatives
}
\description{
Emissivity from atmosphere
}
\references{
Idso SB (1981) A set of equations for full spectrum and 8- to 14-um
 and 10.5- to 12.5-um thermal radiation from cloudless skies.
 Water Resour Res 17:295–304
}

