% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{EID}
\alias{EID}
\title{Emissivity from atmosphere}
\usage{
EID(data, func)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}
}
\value{
a vector with emissivity estimatives
}
\description{
Emissivity from atmosphere
}
\author{
Roilan Hernandez, Guilherme Goergen and Jonatan Tatsch
}

