% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{EKZ}
\alias{EKZ}
\title{Emissivity from atmosphere}
\usage{
EKZ(data, func = "-", coef1 = 0.23, coef2 = 0.484, coef3 = 1/8,
  coef4 = 0.22, coef5 = 1, adjust = FALSE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1, coef2, coef3, coef4, coef5}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}
}
\value{
a vector with emissivity estimatives
}
\description{
Emissivity from atmosphere
}
\references{
Konzelmann T, van de Wal RSW, Greuell W, Bintanja R, Henneken
EAC, Abe-Ouchi A (1994) Parameterization of global and
longwave incoming radiation for the Greenland Ice Sheet. Glob
Planet Chang 9:143–164
}

