% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{EKZ}
\alias{EKZ}
\title{Emissivity from atmosphere}
\usage{
EKZ(data, func)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}
}
\value{
a vector with emissivity estimatives
}
\description{
Emissivity from atmosphere
}

