% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{ENM}
\alias{ENM}
\title{Emissivity from atmosphere}
\usage{
ENM(data, func = "-", coef1 = 0.72, coef2 = 0.09, coef3 = 0.06,
  coef4 = 0.22, coef5 = 1, adjust = FALSE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1, coef2, coef3, coef4, coef5}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}
}
\value{
a vector with emissivity estimatives
}
\description{
Emissivity from atmosphere
}
\references{
Niemala, S. (2001) Comparison of surface radiative flux 
parameterizations part I: Longwave radiation. Atmos. Res., 58, 1-18.
}

