% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{EPR}
\alias{EPR}
\title{Emissivity from atmosphere}
\usage{
EPR(data, func = "-", coef1 = 1, coef2 = 1.2, coef3 = 3, coef4 = 0.22,
  coef5 = 1, adjust = FALSE, forced = TRUE)
}
\arguments{
\item{data}{a data frame with all atmospherics variables}

\item{func}{a function for amount of cloud}

\item{coef1, coef2, coef3, coef4, coef5}{Scheme coeficients}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}

\item{forced}{Forced adjust, default TRUE}
}
\value{
a vector with emissivity estimatives
}
\description{
Emissivity from atmosphere
}
\references{
Prata AJ (1996) A new long-wave formula for estimating downward clearsky
radiation at the surface. Q J Roy Meteor Soc 122:1127–1151
}
