% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{eval.params}
\alias{eval.params}
\title{Function for evaluation of all parameterizations}
\usage{
eval.params(data_, lon = -53.76, lat = -29.72, timezone = -4)
}
\arguments{
\item{data_}{A data frame with all atmospherics variables}

\item{lon}{Longitude from local analisys}

\item{lat}{Latitude from local analisys}

\item{timezone}{Local time diference with GMT (-1 for fluxes measurement)}
}
\value{
List with statistical error information between predicted and observed Li
}
\description{
Function for evaluation of all parameterizations
}

