% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2eval.R
\name{get.AllSchems}
\alias{get.AllSchems}
\title{Function to get all scheme calculation, adjust = TRUE make a adjusting NLS.}
\usage{
get.AllSchems(data, Ovrcst_sch = c("CQB", "CKC", "CCB", "CKZ", "CWU", "CJG",
  "CLM", "CFG"), Emiss_sch = c("EAN", "EBR", "EDO", "EGR", "EIJ", "EID",
  "EKZ", "ENM", "EPR", "EST", "ESW", "EAI"), adjust = FALSE,
  log_file = NULL)
}
\arguments{
\item{data}{Data frame with atmospheric variables}

\item{Ovrcst_sch}{Schemes for cloud cover index}

\item{Emiss_sch}{Schemes of atmosphere emissivity}

\item{adjust}{FALSE, TRUE for NLS adjusting}

\item{log_file}{path to a log file}
}
\value{
Data frame with Li observed and all combintions of schemes for calculations of Li
}
\description{
Function to get all scheme calculation, adjust = TRUE make a adjusting NLS.
}
\author{
Roilan Hernandez
}

