% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2eval.R
\name{get.Li}
\alias{get.Li}
\title{Downward Longwave Radiation from a emissivity and a cloud cover schemes}
\usage{
get.Li(data, E_fun = "FHY", C_fun = "CQB")
}
\arguments{
\item{data}{Data frame with column of date (date), temperature (Ta), 
partial vapor pressure (es), potencial radiation (Rpot), relative humidity (rh), 
atennuation index (K)}

\item{E_fun}{Emissivity scheme}

\item{C_fun}{Cloud cover scheme}
}
\value{
Vector with Downward Longwave Radiation time series
}
\description{
Downward Longwave Radiation from a emissivity and a cloud cover schemes
}

