% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2eval.R
\name{get.Li}
\alias{get.Li}
\title{Downward Longwave Radiation from a emissivity and a cloud cover schemes}
\usage{
get.Li(data_, E_fun = "EAN", C_fun = "CQB", adjust = FALSE,
  log_file = NULL)
}
\arguments{
\item{data_}{Data frame with column of date (date), temperature (Ta), 
partial vapor pressure (es), potencial radiation (Rpot), relative humidity (rh), 
atennuation index (K)}

\item{E_fun}{Emissivity scheme}

\item{C_fun}{Cloud cover scheme}

\item{adjust}{FALSE, TRUE if nonlinear least square adjusting wanted}
}
\value{
Vector with Downward Longwave Radiation time series
}
\description{
Downward Longwave Radiation from a emissivity and a cloud cover schemes
}
\examples{
# Downward longwave for Santa Maria site, January and July of  2014
# without adjusting
Li_sm <- get.Li(data = data2,E_fun = "EAN",C_fun = "CQB")
head(Li_sm)
summary(Li_sm)
}

