% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_roli.R
\name{kloudines}
\alias{kloudines}
\title{Incoming solar radiation atenuattion (K)}
\usage{
kloudines(Rg, dates, lon = -53.76, lat = -29.72, timezone = -4)
}
\arguments{
\item{Rg}{global radiation serie.}

\item{dates}{A vector with dates of time serie, same lenght than Rg}

\item{lon}{longitude from local analisys}

\item{lat}{latitude from local analisys}

\item{timezone}{Local time diference with GMT (-1 for fluxes measurement)}
}
\value{
Vector with K index columns
}
\description{
Incoming solar radiation atenuattion (K)
}
\author{
Roilan Hernandez
}

